Update demo for RFSoC4x2

+++ setup.py
@@ -48,20 +48,21 @@ def copy_notebooks():
     dst_nb_dir = os.path.join(board_project_dir)
     copy_tree(src_nb_dir, dst_nb_dir)
 
-#  copy results to python package
-def copy_results():
-    src_at_dir = os.path.join(repo_board_dir, 'results')
-    dst_at_dir = os.path.join(package_name, 'results')
-    copy_tree(src_at_dir, dst_at_dir)
-    data_files.extend(
-        [os.path.join("..", dst_at_dir, f) for f in os.listdir(dst_at_dir)])
+# copy xrfclk file to python package (gen3 devices only)
+def copy_xrfclk():
+    src_at_dir = os.path.join(repo_board_dir, 'xrfclk')
+    if os.path.exists(src_at_dir):
+        dst_at_dir = os.path.join('xrfclk')
+        copy_tree(src_at_dir, dst_at_dir)
+        data_files.extend(
+            [os.path.join("..", dst_at_dir, f) for f in os.listdir(dst_at_dir)])
 
 check_env()
 check_path()
 copy_overlays()
 copy_drivers()
 copy_notebooks()
-copy_results()
+copy_xrfclk()
 
 setup(
     name=package_name,
